Finish generic credential, oid4vci and oid4vp integrations
