+++ example/src/androidTest/java/com/spruceid/mobilesdkexample/ExampleInstrumentedTest.kt
@@ -1,12 +1,10 @@
 package com.spruceid.mobilesdkexample
 
-import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
-
+import androidx.test.platform.app.InstrumentationRegistry
+import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
-
-import org.junit.Assert.*
 
 /**
  * Instrumented test, which will execute on an Android device.
