+++ example/src/main/java/com/spruceid/mobilesdkexample/credentials/ICredentialView.kt
@@ -1,0 +1,17 @@
+package com.spruceid.mobilesdkexample.credentials
+
+import androidx.compose.runtime.Composable
+
+interface ICredentialView {
+    @Composable
+    fun credentialListItem(withOptions: Boolean): Unit
+
+    @Composable
+    fun credentialListItem(): Unit
+
+    @Composable
+    fun credentialDetails(): Unit
+
+    @Composable
+    fun credentialPreviewAndDetails(): Unit
+}
