+++ example/src/main/java/com/spruceid/mobilesdkexample/navigation/SetupNavGraph.kt
@@ -6,12 +6,12 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.navDeepLink
 import com.spruceid.mobilesdkexample.HomeView
+import com.spruceid.mobilesdkexample.credentials.AddToWalletView
 import com.spruceid.mobilesdkexample.verifier.VerifyDLView
 import com.spruceid.mobilesdkexample.verifier.VerifyEAView
 import com.spruceid.mobilesdkexample.verifier.VerifyVCView
 import com.spruceid.mobilesdkexample.verifiersettings.VerifierSettingsHomeView
 import com.spruceid.mobilesdkexample.viewmodels.IRawCredentialsViewModel
-import com.spruceid.mobilesdkexample.wallet.AddToWalletView
 import com.spruceid.mobilesdkexample.wallet.DispatchQRView
 import com.spruceid.mobilesdkexample.wallet.HandleOID4VPView
 import com.spruceid.mobilesdkexample.wallet.OID4VCIView
@@ -57,10 +57,10 @@ fun SetupNavGraph(
         ) { OID4VCIView(navController, rawCredentialsViewModel) }
         composable(
                 route = Screen.HandleOID4VP.route,
-                deepLinks = listOf(navDeepLink { uriPattern = "oid4vp://{url}" })
+                deepLinks = listOf(navDeepLink { uriPattern = "openid4vp://{url}" })
         ) { backStackEntry ->
             val url = backStackEntry.arguments?.getString("url")!!
-            HandleOID4VPView(navController, rawCredentialsViewModel, url)
+            HandleOID4VPView(navController, rawCredentialsViewModel, "openid4vp://$url")
         }
     }
 }
