+++ example/src/main/java/com/spruceid/mobilesdkexample/ui/theme/Color.kt
@@ -29,11 +29,11 @@ val ColorBase800 = Color(0xFF75675C)
 val ColorBlue600 = Color(0xFF2F6AE1)
 val ColorStone50 = Color(0xFFFAFAF9)
 val ColorStone300 = Color(0xFFD6D3D1)
+val ColorStone500 = Color(0xFF78716C)
 val ColorStone600 = Color(0xFF57534E)
 val ColorStone950 = Color(0xFF0C0A09)
 val ColorEmerald900 = Color(0xFF084C3A)
 val ColorRose600 = Color(0xFFE11D48)
 val TextBase = Color(0xFFFBF9F6)
-val TextPrimary = Color(0xFF0C0A09)
 val BgSurfacePureBlue = Color(0xFF2F6AE1)
 val BorderSecondary = Color(0xFFD6D3D1)
